chore: Combine tests into one file
